fix missing endpoint to get secret code
