+++ backend/app.js
@@ -1,7 +1,7 @@
 const express = require("express")
 const cors = require('cors')
 const { PORT } = require("./constants.js")
-const { generateGrid } = require("./utils.js")
+const { generateGrid, getSecretCode } = require("./utils.js")
 
 const app = express()
 
@@ -15,11 +15,17 @@ app.use(
 app.use(express.json())
 
 let bias = ""
+let code = 0
 
 app.get("/grid", (req, res) => {
   const grid = generateGrid(bias)
+  code = getSecretCode(grid)
 
   res.json(JSON.stringify(grid))
+})
+
+app.get("/code", (req, res) => {
+  res.json(JSON.stringify(code))
 })
 
 app.post("/bias", (req, res) => {
