+++ backend/utils.js
@@ -59,10 +59,51 @@ function generateGrid(bias) {
   return grid
 }
 
+function normalizeValue(count) {
+  if(count < 10) return count
+
+  const n = count.toString()
+  let divisor = parseInt(n[0]) + 1
+  
+  while(count % divisor !== 0) {
+    divisor++
+  }
+
+  return count / divisor 
+}
+
+function getSecretCode(grid) {
+  if(grid[0][0] === "") return 0 
+
+  const seconds = new Date().getSeconds().toString().padStart(2, "0")
+
+  const i = parseInt(seconds[0])
+  const j = parseInt(seconds[1])
+  
+  const firstLetter = grid[i][j]
+  const secondLetter = grid[j][i]
+
+  let firstLetterCount = 0
+  let secondLetterCount = 0
+
+  grid.forEach((row) => {
+    row.forEach((letter)=> {
+      if(letter === firstLetter)  return firstLetterCount++
+      if(letter === secondLetter) return secondLetterCount++
+    })
+  })
+
+  firstLetterCount = normalizeValue(firstLetterCount) 
+  secondLetterCount = normalizeValue(secondLetterCount) 
+
+  return firstLetterCount * 10 + secondLetterCount
+}
+
 module.exports = {
  generateRandomNumber, 
  generateRandomChar,
  generateEmptyGrid,
  generateGrid, 
  insertGridBias,
+ getSecretCode
 }
